+++ libs/gesha-core/src/conversions/v3_0/to_rust_type/components/from_schemas/post_processor/expand_inline_schemas.rs
@@ -1,13 +1,14 @@
-use crate::conversions::v3_0::to_rust_type::from_schemas::to_field_shapes::to_field_shapes;
-use crate::conversions::v3_0::to_rust_type::from_schemas::DefinitionShape::Mod;
-use crate::conversions::v3_0::to_rust_type::from_schemas::TypeShape::{Expanded, Higher};
-use crate::conversions::v3_0::to_rust_type::from_schemas::{
+use crate::conversions::v3_0::to_rust_type::components::from_schemas::to_field_shapes::to_field_shapes;
+use crate::conversions::v3_0::to_rust_type::components::from_schemas::DefinitionShape::Mod;
+use crate::conversions::v3_0::to_rust_type::components::from_schemas::TypeShape::{
+    Array, Expanded, Fixed, Higher, InlineObject, Ref,
+};
+use crate::conversions::v3_0::to_rust_type::components::from_schemas::{
     AllOfItemShape, AllOfShape, DefinitionShape, FieldShape, ModShape, PostProcessor, StructShape,
-    TypeHeaderShape, TypePath, TypeShape,
+    TypeHeaderShape, TypePath,
 };
 use crate::conversions::Result;
 use std::ops::Not;
-use TypeShape::{Array, Fixed, InlineObject, Ref};
 
 impl PostProcessor {
     pub fn process_inline_schemas(
