+++ libs/gesha-core/src/conversions/v3_0/to_rust_type/components/from_schemas/type_header_shape.rs
